Move all interactions into a single test file With the previous approach, the order of the interactions in the contract file could be random. Now the order of the interactions in the contract file is determined by the order of the test cases.
