+++ ui/pact/consumer.jest.config.js
@@ -4,7 +4,6 @@ const regularJestConfig = require('../jest.config')
 module.exports = {
   ...regularJestConfig,
   rootDir: '..',
-  globalSetup: './pact/deleteContracts.ts',
   testMatch: ['**/*.consumer.pact.test.(js|jsx|ts|tsx)'],
   testPathIgnorePatterns: defaults.testPathIgnorePatterns,
   testTimeout: 30000
